requirements.txt
tensorflow>=2.13,<2.17
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
opencv-python-headless>=4.8
Pillow>=9.5
streamlit>=1.28
torch; platform_system=="Linux" # optional, not required
torchvision; platform_system=="Linux" # optional, not required
tensorflow-metal; sys_platform == 'darwin'  # accelerate on Mac Apple Silicon

